requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
websockets==12.0
python-dotenv==1.0.0
gunicorn==21.2.0 